tidy up the code for ABookCore
